fixes packaing to avoid split package problem with jpms
