+++ integration/src/main/java/com/arcadedb/integration/importer/SourceDiscovery.java
@@ -228,32 +228,31 @@ public class SourceDiscovery {
       if (knownFileType.equalsIgnoreCase("csv")) {
         settings.options.put("delimiter", knownDelimiter);
         return new CSVImporterFormat();
-      } else if (knownFileType.equalsIgnoreCase("json"))
+      } else if (knownFileType.equalsIgnoreCase("json")) {
         return new JSONImporterFormat();
-      else if (knownFileType.equalsIgnoreCase("xml"))
+      } else if (knownFileType.equalsIgnoreCase("xml")) {
         return new XMLImporterFormat();
-      else if (knownFileType.equalsIgnoreCase("graphml")){
+      } else if (knownFileType.equalsIgnoreCase("graphml")){
 
         try {
-          final Class<FormatImporter> clazz = (Class<FormatImporter>) Class.forName("com.arcadedb.graph.importer.format.GraphMLImporterFormat");
+          final Class<FormatImporter> clazz = (Class<FormatImporter>) Class.forName("com.arcadedb.gremlin.integration.importer.format.GraphMLImporterFormat");
           return clazz.getConstructor().newInstance();
         } catch (ClassNotFoundException | InvocationTargetException | InstantiationException | IllegalAccessException | NoSuchMethodException e) {
           LogManager.instance().log(this, Level.SEVERE, "Impossible to find importer for 'graphml' ", e);
-
         }
 
-      }
-      else if (knownFileType.equalsIgnoreCase("graphson")){
+      } else if (knownFileType.equalsIgnoreCase("graphson")){
+
         try {
-          final Class<FormatImporter> clazz = (Class<FormatImporter>) Class.forName("com.arcadedb.graph.importer.format.GraphSONImporterFormat");
+          final Class<FormatImporter> clazz = (Class<FormatImporter>) Class.forName("com.arcadedb.gremlin.integration.importer.format.GraphSONImporterFormat");
           return clazz.getConstructor().newInstance();
         } catch (ClassNotFoundException | InvocationTargetException | InstantiationException | IllegalAccessException | NoSuchMethodException e) {
-
           LogManager.instance().log(this, Level.SEVERE, "Impossible to find importer for 'graphson' ", e);
         }
-      }
-      else
+
+      } else {
         LogManager.instance().log(this, Level.WARNING, "File type '%s' is not supported. Trying to understand file type...", knownFileType);
+      }
     }
 
     parser.nextChar();
